Adds Home action to the web app
We want the homepage to be a dashboard of teammates. Right now we've got
4 feature tests to make pass, and the first one does now. This test is
essentially: Teammates exist in the database, can you fetch them
alphaspaghettically and stick 'em in a list?

There are pending specs to complete, to drive us to the teammates and
meetings etc

A note is we have no idea what we're doing, and are wrasslin' with
Hanami as we go.