Adds the ability to create and remove teammates
Teammates are the primary concept in the application; without them,
meetings are pointless?

So - we need to be able to handle adding new ones, and being able to say
goodbye when folks move on to other teams.

Lots not handled here: no unit tests for actions or repos, and
practically no error handling in the actions either. Messaging leaves a
lot to be desired, as does the UX.