Never call a thing "the easy thing"
We added the delete button to our show page and in the course of writing
and running the feature specs discovered an interesting bug we'd written
in - if there were more than 1 teammates in the db then the update stopped
working because as originally written, our update action returned the number
of lines updated rather than the updated object.

We resolved that issue and bravely trudged on. You can now add, edit and delete
a teammate from the db.

Thank you hackdays.